refactor query parser
